pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "synthforge"
version = "0.1.0"
description = "Extensible synthetic data generation and curation toolkit"
authors = [{ name = "You", email = "you@example.com" }]
dependencies = [
  "typer[all]>=0.9.0",
  "pyyaml>=6.0.0",
  "requests>=2.31.0",
  "tqdm>=4.66.0",
  "python-dotenv>=1.0.0",
  "pdfminer.six>=20221105"
]
requires-python = ">=3.10"

[project.scripts]
synthforge = "synthforge.cli:app"
